이재인 : 0814 플레이어 구현 2.

+++ GameEngineCore/GameEngineFolderTexture.h
@@ -32,7 +32,7 @@ public:
 		return textures_.size();
 	}
 
-	GameEngineTexture* GetTexture(int _index)
+	GameEngineTexture* GetTexture(size_t _index)
 	{
 		if (textures_.size() <= _index || 0 > _index)
 		{
